seperated tests dirs, split conftests
