+++ python/tests/async_tests/test_auth.py
@@ -8,15 +8,16 @@ from glide.config import ProtocolVersion
 from glide.constants import OK
 from glide.exceptions import RequestError
 from glide.glide_client import TGlideClient
-from tests.conftest import (
-    NEW_PASSWORD,
-    USERNAME,
-    WRONG_PASSWORD,
+
+from tests.async_tests.conftest import (
     auth_client,
     config_set_new_password,
     kill_connections,
 )
 from tests.utils.utils import (
+    NEW_PASSWORD,
+    USERNAME,
+    WRONG_PASSWORD,
     delete_acl_username_and_password,
     set_new_acl_username_with_password,
 )
