+++ python/tests/async_tests/test_pubsub.py
@@ -17,7 +17,8 @@ from glide.config import (
 from glide.constants import OK
 from glide.exceptions import ConfigurationError
 from glide.glide_client import GlideClient, GlideClusterClient, TGlideClient
-from tests.conftest import create_client
+
+from tests.async_tests.conftest import create_client
 from tests.utils.utils import check_if_server_version_lt, get_random_string
