+++ python/tests/async_tests/test_scan.py
@@ -9,7 +9,8 @@ from glide.config import ProtocolVersion
 from glide.exceptions import RequestError
 from glide.glide import ClusterScanCursor
 from glide.glide_client import GlideClient, GlideClusterClient
-from tests.conftest import create_client
+
+from tests.async_tests.conftest import create_client
 from tests.utils.cluster import ValkeyCluster
 from tests.utils.utils import get_random_string
