+++ python/tests/async_tests/tests_server_modules/test_json.py
@@ -21,7 +21,8 @@ from glide.config import ProtocolVersion
 from glide.constants import OK
 from glide.exceptions import RequestError
 from glide.glide_client import GlideClusterClient, TGlideClient
-from tests.test_async_client import get_random_string
+
+from tests.async_tests.test_async_client import get_random_string
 
 
 def get_random_value(value_type="str"):
