+++ python/tests/conftest.py
@@ -1,41 +1,16 @@
 # Copyright Valkey GLIDE Project Contributors - SPDX Identifier: Apache-2.0
 
-import random
 import sys
-from typing import AsyncGenerator, Generator, List, Optional, Union
+from typing import List
 
 import pytest
 
-from glide.config import (
-    AdvancedGlideClientConfiguration,
-    AdvancedGlideClusterClientConfiguration,
-    BackoffStrategy,
-    GlideClientConfiguration,
-    GlideClusterClientConfiguration,
-    NodeAddress,
-    ProtocolVersion,
-    ReadFrom,
-    ServerCredentials,
-)
-from glide.exceptions import ClosingError
-from glide.glide_client import GlideClient, GlideClusterClient, TGlideClient
-from glide.logger import Level as logLevel
-from glide.logger import Logger
-from glide.routes import AllNodes
-from glide.sync import GlideClient as SyncGlideClient
-from glide.sync import GlideClusterClient as SyncGlideClusterClient
-from glide.sync import TGlideClient as TSyncGlideClient
+from tests.sync_tests.conftest import create_sync_client
 from tests.utils.cluster import ValkeyCluster
-from tests.utils.utils import (
-    set_new_acl_username_with_password,
-    sync_check_if_server_version_lt,
-)
+from tests.utils.utils import sync_check_if_server_version_lt
 
 DEFAULT_HOST = "localhost"
 DEFAULT_PORT = 6379
-DEFAULT_TEST_LOG_LEVEL = logLevel.OFF
-
-Logger.set_logger_config(DEFAULT_TEST_LOG_LEVEL)
 
 
 def pytest_addoption(parser):
@@ -266,388 +241,6 @@ def anyio_backend(request):
         )
 
     return request.param[0]
-
-
-@pytest.fixture(scope="function")
-async def glide_client(
-    request,
-    cluster_mode: bool,
-    protocol: ProtocolVersion,
-) -> AsyncGenerator[TGlideClient, None]:
-    "Get async socket client for tests"
-    client = await create_client(
-        request, cluster_mode, protocol=protocol, request_timeout=5000
-    )
-    yield client
-    await test_teardown(request, cluster_mode, protocol)
-    await client.close()
-
-
-@pytest.fixture(scope="function")
-def glide_sync_client(
-    request,
-    cluster_mode: bool,
-    protocol: ProtocolVersion,
-) -> Generator[TSyncGlideClient, None, None]:
-    "Get async socket client for tests"
-    client = create_sync_client(request, cluster_mode, protocol=protocol)
-    yield client
-    sync_test_teardown(request, cluster_mode, protocol)
-    client.close()
-
-
-@pytest.fixture(scope="function")
-async def management_client(
-    request,
-    cluster_mode: bool,
-    protocol: ProtocolVersion,
-) -> AsyncGenerator[TGlideClient, None]:
-    "Get async socket client for tests, used to manage the state when tests are on the client ability to connect"
-    client = await create_client(request, cluster_mode, protocol=protocol)
-    yield client
-    await test_teardown(request, cluster_mode, protocol)
-    await client.close()
-
-
-@pytest.fixture(scope="function")
-async def acl_glide_client(
-    request,
-    cluster_mode: bool,
-    protocol: ProtocolVersion,
-    management_client: TGlideClient,
-) -> AsyncGenerator[TGlideClient, None]:
-    """
-    Client fot tests that use a server pre-configured with an ACL user.
-    This function first uses the management client to register the USERNAME with INITIAL_PASSWORD,so that
-    the client would be ablt to connect.
-    It then returns a client with this USERNAME and INITIAL_PASSWORD already set as its ServerCredentials.
-    """
-
-    await set_new_acl_username_with_password(
-        management_client, USERNAME, INITIAL_PASSWORD
-    )
-
-    client = await create_client(
-        request,
-        cluster_mode,
-        protocol=protocol,
-        credentials=ServerCredentials(username=USERNAME, password=INITIAL_PASSWORD),
-        request_timeout=2000,
-    )
-    yield client
-    await test_teardown(request, cluster_mode, protocol)
-    await client.close()
-
-
-def create_client_config(
-    request,
-    cluster_mode: bool,
-    credentials: Optional[ServerCredentials] = None,
-    database_id: int = 0,
-    addresses: Optional[List[NodeAddress]] = None,
-    client_name: Optional[str] = None,
-    protocol: ProtocolVersion = ProtocolVersion.RESP3,
-    timeout: Optional[int] = 1000,
-    connection_timeout: Optional[int] = 1000,
-    cluster_mode_pubsub: Optional[
-        GlideClusterClientConfiguration.PubSubSubscriptions
-    ] = None,
-    standalone_mode_pubsub: Optional[
-        GlideClientConfiguration.PubSubSubscriptions
-    ] = None,
-    inflight_requests_limit: Optional[int] = None,
-    read_from: ReadFrom = ReadFrom.PRIMARY,
-    client_az: Optional[str] = None,
-    reconnect_strategy: Optional[BackoffStrategy] = None,
-    valkey_cluster: Optional[ValkeyCluster] = None,
-) -> Union[GlideClusterClientConfiguration, GlideClientConfiguration]:
-    use_tls = request.config.getoption("--tls")
-    if cluster_mode:
-        valkey_cluster = valkey_cluster or pytest.valkey_cluster  # type: ignore
-        assert type(valkey_cluster) is ValkeyCluster
-        assert database_id == 0
-        k = min(3, len(valkey_cluster.nodes_addr))
-        seed_nodes = random.sample(valkey_cluster.nodes_addr, k=k)
-        config = GlideClusterClientConfiguration(
-            addresses=seed_nodes if addresses is None else addresses,
-            use_tls=use_tls,
-            credentials=credentials,
-            client_name=client_name,
-            protocol=protocol,
-            request_timeout=timeout,
-            pubsub_subscriptions=cluster_mode_pubsub,
-            inflight_requests_limit=inflight_requests_limit,
-            read_from=read_from,
-            client_az=client_az,
-            advanced_config=AdvancedGlideClusterClientConfiguration(connection_timeout),
-        )
-    else:
-        assert type(pytest.standalone_cluster) is ValkeyCluster  # type: ignore
-        config = GlideClientConfiguration(
-            addresses=(
-                pytest.standalone_cluster.nodes_addr if addresses is None else addresses  # type: ignore
-            ),
-            use_tls=use_tls,
-            credentials=credentials,
-            database_id=database_id,
-            client_name=client_name,
-            protocol=protocol,
-            request_timeout=timeout,
-            pubsub_subscriptions=standalone_mode_pubsub,
-            inflight_requests_limit=inflight_requests_limit,
-            read_from=read_from,
-            client_az=client_az,
-            reconnect_strategy=reconnect_strategy,
-            advanced_config=AdvancedGlideClientConfiguration(connection_timeout),
-        )
-    return config
-
-
-async def create_client(
-    request,
-    cluster_mode: bool,
-    credentials: Optional[ServerCredentials] = None,
-    database_id: int = 0,
-    addresses: Optional[List[NodeAddress]] = None,
-    client_name: Optional[str] = None,
-    protocol: ProtocolVersion = ProtocolVersion.RESP3,
-    request_timeout: Optional[int] = 1000,
-    connection_timeout: Optional[int] = 1000,
-    cluster_mode_pubsub: Optional[
-        GlideClusterClientConfiguration.PubSubSubscriptions
-    ] = None,
-    standalone_mode_pubsub: Optional[
-        GlideClientConfiguration.PubSubSubscriptions
-    ] = None,
-    inflight_requests_limit: Optional[int] = None,
-    read_from: ReadFrom = ReadFrom.PRIMARY,
-    client_az: Optional[str] = None,
-    reconnect_strategy: Optional[BackoffStrategy] = None,
-    valkey_cluster: Optional[ValkeyCluster] = None,
-) -> Union[GlideClient, GlideClusterClient]:
-    config = create_client_config(
-        request,
-        cluster_mode,
-        credentials,
-        database_id,
-        addresses,
-        client_name,
-        protocol,
-        request_timeout,
-        connection_timeout,
-        cluster_mode_pubsub,
-        standalone_mode_pubsub,
-        inflight_requests_limit,
-        read_from,
-        client_az,
-        reconnect_strategy,
-        valkey_cluster,
-    )
-    if cluster_mode:
-        return await GlideClusterClient.create(config)
-    else:
-        return await GlideClient.create(config)
-
-
-def create_sync_client(
-    request,
-    cluster_mode: bool,
-    credentials: Optional[ServerCredentials] = None,
-    database_id: int = 0,
-    addresses: Optional[List[NodeAddress]] = None,
-    client_name: Optional[str] = None,
-    protocol: ProtocolVersion = ProtocolVersion.RESP3,
-    timeout: Optional[int] = 1000,
-    connection_timeout: Optional[int] = 1000,
-    cluster_mode_pubsub: Optional[
-        GlideClusterClientConfiguration.PubSubSubscriptions
-    ] = None,
-    standalone_mode_pubsub: Optional[
-        GlideClientConfiguration.PubSubSubscriptions
-    ] = None,
-    inflight_requests_limit: Optional[int] = None,
-    read_from: ReadFrom = ReadFrom.PRIMARY,
-    client_az: Optional[str] = None,
-    reconnect_strategy: Optional[BackoffStrategy] = None,
-    valkey_cluster: Optional[ValkeyCluster] = None,
-) -> TSyncGlideClient:
-    # Create sync client
-    config = create_client_config(
-        request,
-        cluster_mode,
-        credentials,
-        database_id,
-        addresses,
-        client_name,
-        protocol,
-        timeout,
-        connection_timeout,
-        cluster_mode_pubsub,
-        standalone_mode_pubsub,
-        inflight_requests_limit,
-        read_from,
-        client_az,
-        reconnect_strategy,
-        valkey_cluster,
-    )
-    if cluster_mode:
-        return SyncGlideClusterClient.create(config)
-    else:
-        return SyncGlideClient.create(config)
-
-
-USERNAME = "username"
-INITIAL_PASSWORD = "initial_password"
-NEW_PASSWORD = "new_secure_password"
-WRONG_PASSWORD = "wrong_password"
-
-
-async def auth_client(client: TGlideClient, password: str, username: str = "default"):
-    """
-    Authenticates the given TGlideClient server connected. If no username is provided, uses the 'default' user.
-    """
-    if isinstance(client, GlideClient):
-        return await client.custom_command(["AUTH", username, password])
-    elif isinstance(client, GlideClusterClient):
-        return await client.custom_command(
-            ["AUTH", username, password], route=AllNodes()
-        )
-
-
-def sync_auth_client(client: TSyncGlideClient, password):
-    """
-    Authenticates the given TGlideClient server connected.
-    """
-    if isinstance(client, GlideClient):
-        client.custom_command(["AUTH", password])
-    elif isinstance(client, GlideClusterClient):
-        client.custom_command(["AUTH", password], route=AllNodes())
-
-
-def sync_config_set_new_password(client: TSyncGlideClient, password):
-    """
-    Sets a new password for the given TGlideClient server connected.
-    This function updates the server to require a new password.
-    """
-    if isinstance(client, GlideClient):
-        client.config_set({"requirepass": password})
-    elif isinstance(client, GlideClusterClient):
-        client.config_set({"requirepass": password}, route=AllNodes())
-
-
-async def config_set_new_password(client: TGlideClient, password):
-    """
-    Sets a new password for the given TGlideClient server connected.
-    This function updates the server to require a new password.
-    """
-    if isinstance(client, GlideClient):
-        await client.config_set({"requirepass": password})
-    elif isinstance(client, GlideClusterClient):
-        await client.config_set({"requirepass": password}, route=AllNodes())
-
-
-def sync_kill_connections(client: TSyncGlideClient):
-    """
-    Kills all connections to the given TGlideClient server connected.
-    """
-    if isinstance(client, GlideClient):
-        client.custom_command(["CLIENT", "KILL", "TYPE", "normal"])
-    elif isinstance(client, GlideClusterClient):
-        client.custom_command(["CLIENT", "KILL", "TYPE", "normal"], route=AllNodes())
-
-
-async def kill_connections(client: TGlideClient):
-    """
-    Kills all connections to the given TGlideClient server connected.
-    """
-    if isinstance(client, GlideClient):
-        await client.custom_command(["CLIENT", "KILL", "TYPE", "normal"])
-    elif isinstance(client, GlideClusterClient):
-        await client.custom_command(
-            ["CLIENT", "KILL", "TYPE", "normal"], route=AllNodes()
-        )
-
-
-async def test_teardown(request, cluster_mode: bool, protocol: ProtocolVersion):
-    """
-    Perform teardown tasks such as flushing all data from the cluster.
-
-    If authentication is required, attempt to connect with the known password,
-    reset it back to empty, and proceed with teardown.
-    """
-    credentials = None
-    try:
-        # Try connecting without credentials
-        client = await create_client(
-            request, cluster_mode, protocol=protocol, request_timeout=2000
-        )
-        await client.custom_command(["FLUSHALL"])
-        await client.close()
-    except ClosingError as e:
-        # Check if the error is due to authentication
-        if "NOAUTH" in str(e):
-            # Use the known password to authenticate
-            credentials = ServerCredentials(password=NEW_PASSWORD)
-            client = await create_client(
-                request,
-                cluster_mode,
-                protocol=protocol,
-                request_timeout=2000,
-                credentials=credentials,
-            )
-            try:
-                await auth_client(client, NEW_PASSWORD)
-                # Reset the server password back to empty
-                await config_set_new_password(client, "")
-                await client.update_connection_password(None)
-                # Perform the teardown
-                await client.custom_command(["FLUSHALL"])
-            finally:
-                await client.close()
-        else:
-            raise e
-
-
-def sync_test_teardown(request, cluster_mode: bool, protocol: ProtocolVersion):
-    """
-    Perform teardown tasks such as flushing all data from the cluster.
-
-    If authentication is required, attempt to connect with the known password,
-    reset it back to empty, and proceed with teardown.
-    """
-    credentials = None
-    try:
-        # Try connecting without credentials
-        client = create_sync_client(
-            request, cluster_mode, protocol=protocol, timeout=2000
-        )
-        client.custom_command(["FLUSHALL"])
-        client.close()
-    except ClosingError as e:
-        # Check if the error is due to authentication
-        if "NOAUTH" in str(e):
-            # Use the known password to authenticate
-            credentials = ServerCredentials(password=NEW_PASSWORD)
-            client = create_sync_client(
-                request,
-                cluster_mode,
-                protocol=protocol,
-                timeout=2000,
-                credentials=credentials,
-            )
-            try:
-                sync_auth_client(client, NEW_PASSWORD)
-                # Reset the server password back to empty
-                sync_config_set_new_password(client, "")
-                client.update_connection_password(None)
-                # Perform the teardown
-                client.custom_command(["FLUSHALL"])
-            finally:
-                client.close()
-        else:
-            raise e
-
-
 @pytest.fixture(autouse=True)
 def skip_if_version_below(request):
     """
@@ -661,8 +254,11 @@ def skip_if_version_below(request):
     if request.node.get_closest_marker("skip_if_version_below"):
         min_version = request.node.get_closest_marker("skip_if_version_below").args[0]
         client = create_sync_client(request, False)
-        if sync_check_if_server_version_lt(client, min_version):
-            pytest.skip(
-                reason=f"This feature added in version {min_version}",
-                allow_module_level=True,
-            )
+        try: 
+            if sync_check_if_server_version_lt(client, min_version):
+                pytest.skip(
+                    reason=f"This feature added in version {min_version}",
+                    allow_module_level=True,
+                )
+        finally:
+            client.close()
